// src/types/types.tsx
export interface Starship {
  MGLT: string;
  cargo_capacity: string;
  consumables: string;
  crew: string;
  cost_in_credits: string;
  hyperdrive_rating: string;
  length: string;
  manufacturer: string;
  max_atmosphering_speed: string;
  model: string;
  name: string;
  passengers: string;
  starship_class: string;
}
export interface Character {
  birth_year: string;
  eye_color: string;
  gender: string;
  hair_color: string;
  height: string;
  mass: string;
  name: string;
  skin_color: string;
}